Refactor stock models and APIs to separate finance and market-related components
