+++ library/src/commonMain/kotlin/li/mercury/tushare/TuShare.kt
@@ -15,8 +15,6 @@ import li.mercury.tushare.api.index.IndexApi
 import li.mercury.tushare.api.index.IndexApiInterface
 import li.mercury.tushare.api.news.NewsApi
 import li.mercury.tushare.api.news.NewsApiInterface
-import li.mercury.tushare.api.stock.StockApi
-import li.mercury.tushare.api.stock.StockApiInterface
 import li.mercury.tushare.api.stock.basic.StockBasicApi
 import li.mercury.tushare.api.stock.basic.StockBasicApiInterface
 import li.mercury.tushare.api.stock.character.StockCharacterApi
@@ -65,9 +63,6 @@ class TuShare(
 
     /** 新闻相关API */
     val news: NewsApiInterface by lazy { NewsApi(this) }
-
-    /** 股票信息相关API */
-    val stock: StockApiInterface by lazy { StockApi(this) }
 
     /** 股票基础信息相关API */
     val stockBasic: StockBasicApiInterface by lazy { StockBasicApi(this) }
