+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/basic/StockBasicApiInterface.kt
@@ -19,8 +19,8 @@ import li.mercury.tushare.api.stock.basic.models.StockCompanyParams
 import li.mercury.tushare.api.stock.basic.models.StockCompanyResult
 import li.mercury.tushare.api.stock.basic.models.TradeCalParams
 import li.mercury.tushare.api.stock.basic.models.TradeCalResult
-import li.mercury.tushare.api.stock.models.StkPremarketParams
-import li.mercury.tushare.api.stock.models.StkPremarketResult
+import li.mercury.tushare.api.stock.basic.models.StkPremarketParams
+import li.mercury.tushare.api.stock.basic.models.StkPremarketResult
 
 /**
  * 股票相关API的存储库接口
