+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/basic/models/HsConstParams.kt
@@ -2,7 +2,7 @@ package li.mercury.tushare.api.stock.basic.models
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import li.mercury.tushare.api.stock.models.HsType
+import li.mercury.tushare.api.stock.basic.models.HsType
 
 /**
  * 沪深股通成分股查询参数
