+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/basic/models/HsTarget.kt
@@ -1,4 +1,4 @@
-package li.mercury.tushare.api.stock.models
+package li.mercury.tushare.api.stock.basic.models
 
 import kotlinx.serialization.Serializable
