+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/basic/models/StockBasicParams.kt
@@ -2,8 +2,8 @@ package li.mercury.tushare.api.stock.basic.models
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import li.mercury.tushare.api.stock.models.HsTarget
-import li.mercury.tushare.api.stock.models.ListStatus
+import li.mercury.tushare.api.stock.basic.models.HsTarget
+import li.mercury.tushare.api.stock.basic.models.ListStatus
 import li.mercury.tushare.models.Exchange
 import li.mercury.tushare.models.TsCode
