+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/basic/models/StockBasicResult.kt
@@ -3,8 +3,8 @@ package li.mercury.tushare.api.stock.basic.models
 import kotlinx.datetime.LocalDate
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import li.mercury.tushare.api.stock.models.HsTarget
-import li.mercury.tushare.api.stock.models.ListStatus
+import li.mercury.tushare.api.stock.basic.models.HsTarget
+import li.mercury.tushare.api.stock.basic.models.ListStatus
 import li.mercury.tushare.models.Exchange
 import li.mercury.tushare.models.TsCode
 import li.mercury.tushare.utils.LocalDateAsStringSerializer
