+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/finance/models/ReportType.kt
@@ -1,4 +1,4 @@
-package li.mercury.tushare.api.stock.models
+package li.mercury.tushare.api.stock.finance.models
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
