+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/market/models/HsMarketType.kt
@@ -1,3 +1,5 @@
+package li.mercury.tushare.api.stock.market.models
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
