+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/market/models/HsgtTop10Params.kt
@@ -1,6 +1,6 @@
 package li.mercury.tushare.api.stock.market.models
 
-import HsMarketType
+import li.mercury.tushare.api.stock.market.models.HsMarketType
 import kotlinx.datetime.LocalDate
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
