+++ library/src/commonMain/kotlin/li/mercury/tushare/api/stock/market/models/SuspendDParams.kt
@@ -3,7 +3,7 @@ package li.mercury.tushare.api.stock.market.models
 import kotlinx.datetime.LocalDate
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import li.mercury.tushare.api.stock.models.SuspendType
+import li.mercury.tushare.api.stock.market.models.SuspendType
 import li.mercury.tushare.models.TsCode
 import li.mercury.tushare.utils.LocalDateAsStringSerializer
