+++ library/src/commonTest/kotlin/li/mercury/tushare/StockBasicApiTest.kt
@@ -16,8 +16,8 @@ import li.mercury.tushare.api.stock.basic.models.StkManagersParams
 import li.mercury.tushare.api.stock.basic.models.StkRewardsParams
 import li.mercury.tushare.api.stock.basic.models.StockCompanyParams
 import li.mercury.tushare.api.stock.basic.models.TradeCalParams
-import li.mercury.tushare.api.stock.models.HsType
-import li.mercury.tushare.api.stock.models.StkPremarketParams
+import li.mercury.tushare.api.stock.basic.models.HsType
+import li.mercury.tushare.api.stock.basic.models.StkPremarketParams
 import li.mercury.tushare.models.Exchange
 import li.mercury.tushare.models.TsCode
 import okio.FileSystem
